requirements.txt
mysql-connector-python==8.0.25
protobuf==3.17.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.4
